+++ Source/Mosa.Korlib/System/Int64.cs
@@ -68,12 +68,12 @@ namespace System
 
 			for (int i = z; i < s.Length; i++)
 			{
-				int ind = (int)digits.IndexOf(s[i]);
+				int ind = digits.IndexOf(s[i]);
 				if (ind == -1)
 				{
 					throw new Exception("Format is incorrect");
 				}
-				result = (long)((result * 10) + ind);
+				result = (result * 10) + ind;
 			}
 
 			if (neg) result *= -1;
